backend/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
diffusers==0.25.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
ollama==0.5.3
omegaconf==2.3.0
packaging<26.0.0
pillow==11.0.0
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.3
regex==2025.9.18
reportlab
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.3
tokenizers==0.15.2
# torch==2.9.1+cu128
# torchvision==0.24.1+cu128
# NOTE: Install torch manually with CUDA support:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
tqdm==4.67.1
transformers==4.37.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
email-validator==2.2.0

# IMPORTANT:
# Do NOT pin torch here. Install it manually to match your CUDA.
# Example:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121

# Face Recognition & Image Processing
# NOTE: Swapped opencv-python -> opencv-python-headless (no GUI needed on server)
#       opencv-python-headless and opencv-python conflict (same cv2 module)
insightface==0.7.3
onnxruntime-gpu==1.16.0
opencv-python-headless  # Replaces opencv-python; install insightface first

# Testing Dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.0
langchain
langchain-community
piper-tts
pathvalidate  # Required by piper-tts
onnxruntime

# Speech-to-Text (Offline)
openai-whisper

# Story & Image Evaluation
sentence-transformers==2.6.1  # Text coherence analysis
textstat  # Readability metrics
